requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.21.2
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
PyJWT==1.7.1
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
zipp==3.15.0
